requirements.txt
python_twitter==3.5
Requests==2.32.3